refactor(math): consolidate polynomial family under polynomials
Merge the ticket-shaped singular polynomial operation package into the
surviving polynomials owner and fold the multivariate polynomial,
polynomial map, and real algebra domains in as subpackages. Operation
IDs and request/result schemas are unchanged.

- polynomial private operation modules -> polynomials/
- multivariate_polynomial -> polynomials/multivariate
- polynomial_maps -> polynomials/maps
- real_algebra -> polynomials/real_algebra

Kept standalone: formal_power_series (distinct value type),
root_isolation (self-contained real-root kernels), and modular_polynomials
(lone native module with no shared-value dependency).

Validation: make check green (ruff, mypy, 977 tests); import-contracts
and architecture checks pass; operation-schema snapshot unchanged at 360.
