--- src/jacobian/math/polynomial/__init__.py
@@ -1,3 +1,0 @@
-"""Polynomial operation declaration ownership."""
-
-__all__: list[str] = []
