+++ src/jacobian/math/polynomials/_conversions.py
