+++ src/jacobian/math/polynomials/_models.py
