+++ src/jacobian/math/polynomials/_support.py
