+++ src/jacobian/math/polynomials/_syzygy_models.py
