+++ src/jacobian/math/polynomials/maps/_models.py
