+++ src/jacobian/math/polynomials/multivariate/__init__.py
