+++ src/jacobian/math/polynomials/multivariate/_models.py
