+++ src/jacobian/math/polynomials/real_algebra/__init__.py
@@ -1,0 +1,5 @@
+"""Exact real algebra operations."""
+
+from jacobian.math.polynomials.real_algebra.operations import root_count, sturm_chain
+
+__all__ = ["root_count", "sturm_chain"]
