+++ src/jacobian/math/polynomials/real_algebra/_models.py
