+++ src/jacobian/math/polynomials/real_algebra/operations.py
