--- src/jacobian/math/real_algebra/__init__.py
@@ -1,5 +1,0 @@
-"""Exact real algebra operations."""
-
-from jacobian.math.real_algebra.operations import root_count, sturm_chain
-
-__all__ = ["root_count", "sturm_chain"]
