+++ tests/math/polynomials/test_polynomial_conversions.py
@@ -1,11 +1,11 @@
 from __future__ import annotations
 
-from jacobian.math.polynomial._conversions import (
+from jacobian.math.polynomials._conversions import (
     rational_polynomial_from_sympy,
     rational_polynomial_to_sympy,
 )
-from jacobian.math.polynomial._models import PolynomialGcdRequest
-from jacobian.math.polynomial._operations import polynomial_gcd
+from jacobian.math.polynomials._models import PolynomialGcdRequest
+from jacobian.math.polynomials._operations import polynomial_gcd
 from jacobian.math.polynomials.values import (
     RationalPolynomial,
     SparseRationalPolynomial,
