+++ tests/math/polynomials/test_provider_conversions.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import pytest
 from sympy import Poly, symbols
 
-from jacobian.math.polynomial._conversions import rational_polynomial_from_sympy
+from jacobian.math.polynomials._conversions import rational_polynomial_from_sympy
 
 
 def test_sympy_conversion_rejects_non_qq_or_reordered_generators() -> None:
